祖先種に状態遷移を実装
