+++ src/simulations/modular_life_v2/source.ts
@@ -146,7 +146,7 @@ function initializeMaterials(world: World): void {
 }
 
 function initializeAncestors(world: World): void {
-  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.selfReproduction(NeighbourDirections.bottom)), world.size.div(3).floor())
-  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.selfReproduction(NeighbourDirections.right)), world.size.div(2).floor())
-  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.selfReproduction(NeighbourDirections.top)), world.size.div(3).mult(2).floor())
+  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.minimumSelfReproduction(NeighbourDirections.bottom)), world.size.div(3).floor())
+  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.minimumSelfReproduction(NeighbourDirections.right)), world.size.div(2).floor())
+  world.addAncestor(Ancestor.minimumSelfReproduction(AncestorCode.minimumSelfReproduction(NeighbourDirections.top)), world.size.div(3).mult(2).floor())
 }
