+++ src/simulations/modular_life_v2/system.ts
@@ -1,3 +1,3 @@
 export const System = {
-  version: "2.2.7"
+  version: "2.2.8"
 } as const
